scrapy_engine/scrapy_engine/requirements.txt
beautifulsoup4==4.12.3
boto3==1.36.1
huggingface_hub==0.27.1
langid==1.1.6
pymongo==4.6.1
python-dotenv==1.0.1
# redis==5.0.1
requests==2.31.0
Scrapy==2.11.2
Twisted==22.10.0